pyproject.toml
[tool.poetry]
name = "checkbox-test"
version = "0.1.0"
description = ""
authors = ["i_konov <i_konov@ukr.net>"]

[tool.poetry.dev-dependencies]
mypy = "*"
black = "==20.8b1"
isort = "==5.7.0"
pytest = "^7"
pytest-env = "^0.6.2"
ipython = "==7.19.0"
flake8 = "==3.8.4"
coverage = "==5.3.1"
flake8-logging-format = "*"
flake8-alfred = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-coding = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-eradicate = "*"
flake8-multiline-containers = "*"
flake8-pep3101 = "*"
flake8-print = "*"
flake8-pytest-style = "*"
flake8-return = "*"
flake8-quotes = "*"
flake8-simplify = "*"
pep8-naming = "*"
flake8-cognitive-complexity = "*"
flake8-isort = "*"
sqlalchemy-stubs = "^0.4"
flake8-annotations = "^2.6.2"
types-requests = "^2.25.6"
types-aiofiles = "^0.8.4"
flake8-commas = "^2.1.0"
add-trailing-comma = "^2.4.0"
pytest-mock = "^3.11.1"
autoflake = "^1.7.7"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.88.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
uvicorn = "^0.13.4"
requests = "==2.25.1"
sentry-sdk = "^1.12.1"
psutil = "==5.8.0"
SQLAlchemy = "^1.3.23"
asyncpg = "0.27"
psycopg2-binary = "^2.8.6"
structlog = "^21.1.0"
alembic = "^1.5.7"
aiofiles = "^0.6.0"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
gunicorn = "^20.1.0"
structlog-sentry = "^1.4.0"
celery = {version = "4.4.7", extras = ["redis"]}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
Jinja2 = "^3.1.1"
emails = "^0.6"
python-multipart = "^0.0.5"
multipart = "^0.2.4"
SQLAlchemy-Utils = "^0.38.2"
aioredis = "^2.0.1"
openpyxl = "^3.0.10"
python-slugify = "^6.1.2"
beautifulsoup4 = "^4.11.1"
aiosmtplib = "^2.0.0"
pytest-asyncio = "^0.20.2"
fastapi-utils = "^0.2.1"
httpagentparser = "^1.9.5"
transaction = "^3.0.1"
pytest-cov = "^4.0.0"
databases = "^0.6.2"
httpx = "^0.23.1"
phonenumbers = "^8.13.2"
aioshutil = "^1.2"
python-dotenv = "^0.21.0"
argparse = "^1.4.0"
croniter = "^1.3.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
env = "TESTING=True"
minversion = "6.0"
pythonpath = [
  ".", "src"
]
testpaths = [
    "tests",
]

[tool.black]
line-length = 100
py38 = true
skip-string-normalization = 1
include = '\.py$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | env
  | test-reports
  | alembic
  | ptvs_virtualenv_proxy.py
)/
'''

[tool.isort]
line_length = 100
default_section = "FIRSTPARTY"
force_single_line = "True"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_local_folder = "api,core,crud,db,enums,models,schemas,sdk,tests,validators"
skip = [ "alembic" ]
